CONFLICT (content): Merge conflict in openmp/libomptarget/plugins/level0/CMakeLists.txt

  commit 2e0cd2c3268212013a49292abc2da6c8b5f5ead3
  Author: Yilong Guo <vfirst218@gmail.com>
  Date:   Thu May 19 11:26:28 2022 +0800

      Revert "[OpenMP][Offload][L0] Enable sycl wrapper for interop (#4976)" (#5111)

      This reverts commit 9d6bc3664a3aed9e37568dc736adb948f9db8b45.

+++ openmp/libomptarget/include/omptargetplugin.h
@@ -289,7 +289,7 @@ EXTERN void __tgt_rtl_deinit(void);
 // Create OpenMP interop with the given interop context
 EXTERN __tgt_interop *__tgt_rtl_create_interop(
     int32_t ID, int32_t InteropContext, int32_t NumPrefers,
-    int32_t *PreferIDs);
+    intptr_t *PreferIDs);
 
 // Release OpenMP interop
 EXTERN int32_t __tgt_rtl_release_interop(int32_t ID, __tgt_interop *Interop);
